Refactor item class structure.
