+++ src/com/dakoda/alr/item/misc/ItemQuest.java
@@ -1,5 +1,7 @@
-package com.dakoda.alr.item;
+package com.dakoda.alr.item.misc;
 import com.dakoda.alr.game.Quest;
+import com.dakoda.alr.item.Item;
+
 import static com.dakoda.alr.item.ItemType.QUEST;
 
 public class ItemQuest extends Item {
